+++ lua/terrortown/entities/roles/detective/shared.lua
@@ -34,16 +34,13 @@ function ROLE:PreInitialize()
 end
 
 if CLIENT then
-	hook.Add("TTT2ModifyRoleSettingsMenu_Credits", "ttt2_add_role_menu_settings_detective", function(role, roleData, parent)
-		-- only detective (sub-)roles have this convar
-		if role ~= ROLE_DETECTIVE and roleData.baserole ~= ROLE_DETECTIVE then return end
-
+	function ROLE:AddToSettingsMenuCreditsForm(parent)
 		parent:MakeSlider({
-			serverConvar = "ttt_" .. roleData.abbr .. "_credits_traitordead",
+			serverConvar = "ttt_det_credits_traitordead",
 			label = "label_roles_credits_traitordead",
 			min = 0,
 			max = 10,
 			decimal = 0
 		})
-	end)
+	end
 end