+++ lua/terrortown/menus/gamemode/roles/base_roles.lua
@@ -110,10 +110,7 @@ local function PopulateCredits(parent, roleData)
 	})
 
 	-- run a hook to add role specific custom credit convars
-
-	---
-	-- @realm client
-	hook.Run("TTT2ModifyRoleSettingsMenu_Credits", roleData.index, roleData, parent)
+	roleData:AddToSettingsMenuCreditsForm(parent)
 end
 
 function CLGAMEMODESUBMENU:Populate(parent)
@@ -127,8 +124,5 @@ function CLGAMEMODESUBMENU:Populate(parent)
 	PopulateCredits(vgui.CreateTTT2Form(parent, "header_roles_credits"), self.roleData)
 
 	-- run a hook to add role specific custom convars
-
-	---
-	-- @realm client
-	hook.Run("TTT2ModifyRoleSettingsMenu", self.roleData.index, self.roleData, parent)
+	self.roleData:AddToSettingsMenu(parent)
 end